Facebook Rental Search Tool
Languages : Node, Angular, Bootstrap, Html5, Facebook Graph API and
CSS3.
A light weighted tool to easily search and seek rooms for
rent/share/ownership in a particular area. With a way to diffrentiate
and implementation of machine learning to to know the frequency and time
of data posting as well as it's regulation. A on click view to know the
details of various facebook pages and their stats.